# perf(profiling): Remove recursive call in text renderer