The function call overhead in the text renderer can be significant with a large
number of frames. This flattens the recursive procedure to use iteration to
avoid the overhead.

# Before
```
typescript (full profile) x 4,174 ops/sec ±2.03% (74 runs sampled)
typescript (center half) x 8,630 ops/sec ±3.07% (82 runs sampled)
typescript (right quarter) x 10,370 ops/sec ±5.08% (74 runs sampled)
```

# After
```
typescript (full profile) x 6,151 ops/sec ±2.58% (81 runs sampled)
typescript (center half) x 14,881 ops/sec ±4.76% (81 runs sampled)
typescript (right quarter) x 20,735 ops/sec ±2.98% (84 runs sampled)
```